[aGrUM] better bootstrap for EM in DAG2BNLearner

+++ src/testunits/module_BN/DAGModelTestSuite.h
@@ -34,18 +34,18 @@ namespace gum_tests {
   class DAGModelTestSuite : public CxxTest::TestSuite {
     public:
     void testEquality() {
-      auto bn = gum::BayesNet< int >::fastPrototype("a->b->c;a->c");
+      auto bn = gum::BayesNet< float >::fastPrototype("a->b->c;a->c");
 
       TS_ASSERT(bn.hasSameStructure(bn));
       TS_ASSERT(
-         bn.hasSameStructure(gum::BayesNet< int >::fastPrototype("a->b->c;a->c")));
+         bn.hasSameStructure(gum::BayesNet< float >::fastPrototype("a->b->c;a->c")));
       TS_ASSERT(!bn.hasSameStructure(
-         gum::BayesNet< int >::fastPrototype("b->a->c;b->c")));
-      TS_ASSERT(!bn.hasSameStructure(gum::BayesNet< int >::fastPrototype("b->a")));
+         gum::BayesNet< float >::fastPrototype("b->a->c;b->c")));
+      TS_ASSERT(!bn.hasSameStructure(gum::BayesNet< float >::fastPrototype("b->a")));
       TS_ASSERT(
          !bn.hasSameStructure(gum::BayesNet< double >::fastPrototype("b->a")));
       TS_ASSERT(
-         !bn.hasSameStructure(gum::BayesNet< int >::fastPrototype("b->a->d")));
+         !bn.hasSameStructure(gum::BayesNet< float >::fastPrototype("b->a->d")));
     }
   };
 }   // namespace gum_tests
