+++ src/testunits/module_BN/MarkovBlanketTestSuite.h
@@ -34,7 +34,7 @@ namespace gum_tests {
   class MarkovBlanketTestSuite : public CxxTest::TestSuite {
     public:
     void testChain() {
-      auto bn = gum::BayesNet< int >::fastPrototype("a->b->c");
+      auto bn = gum::BayesNet< float >::fastPrototype("a->b->c");
       {
         auto mb = gum::MarkovBlanket(bn, "a");
 
@@ -77,34 +77,34 @@ namespace gum_tests {
 
     void testMarkovBlanketStructure() {
       const auto bn =
-         gum::BayesNet< int >::fastPrototype("a->b->c->d->e;f->d->g;h->i->g");
+         gum::BayesNet< float >::fastPrototype("a->b->c->d->e;f->d->g;h->i->g");
       TS_ASSERT(!gum::MarkovBlanket(bn, "a").hasSameStructure(
-         gum::BayesNet< int >::fastPrototype("b->a")));
+         gum::BayesNet< float >::fastPrototype("b->a")));
 
       TS_ASSERT(gum::MarkovBlanket(bn, "a").hasSameStructure(
-         gum::BayesNet< int >::fastPrototype("a->b")));
+         gum::BayesNet< float >::fastPrototype("a->b")));
       TS_ASSERT(gum::MarkovBlanket(bn, "b").hasSameStructure(
-         gum::BayesNet< int >::fastPrototype("a->b->c")));
+         gum::BayesNet< float >::fastPrototype("a->b->c")));
       TS_ASSERT(gum::MarkovBlanket(bn, "c").hasSameStructure(
-         gum::BayesNet< int >::fastPrototype("b->c->d;f->d")));
+         gum::BayesNet< float >::fastPrototype("b->c->d;f->d")));
       TS_ASSERT(gum::MarkovBlanket(bn, "d").hasSameStructure(
-         gum::BayesNet< int >::fastPrototype("c->d->e;f->d->g;i->g")));
+         gum::BayesNet< float >::fastPrototype("c->d->e;f->d->g;i->g")));
       TS_ASSERT(gum::MarkovBlanket(bn, "e").hasSameStructure(
-         gum::BayesNet< int >::fastPrototype("d->e")));
+         gum::BayesNet< float >::fastPrototype("d->e")));
       TS_ASSERT(gum::MarkovBlanket(bn, "f").hasSameStructure(
-         gum::BayesNet< int >::fastPrototype("c->d;f->d;")));
+         gum::BayesNet< float >::fastPrototype("c->d;f->d;")));
       TS_ASSERT(gum::MarkovBlanket(bn, "g").hasSameStructure(
-         gum::BayesNet< int >::fastPrototype("d->g;i->g;")));
+         gum::BayesNet< float >::fastPrototype("d->g;i->g;")));
       TS_ASSERT(gum::MarkovBlanket(bn, "h").hasSameStructure(
-         gum::BayesNet< int >::fastPrototype("h->i;")));
+         gum::BayesNet< float >::fastPrototype("h->i;")));
       TS_ASSERT(gum::MarkovBlanket(bn, "i").hasSameStructure(
-         gum::BayesNet< int >::fastPrototype("d->g;h->i->g;;")));
+         gum::BayesNet< float >::fastPrototype("d->g;h->i->g;;")));
     }
 
     void testMarkovBlanketSpecialArcs() {
-      const auto bn = gum::BayesNet< int >::fastPrototype(
+      const auto bn = gum::BayesNet< float >::fastPrototype(
          "aa->bb->cc->dd->ee;ff->dd->gg;hh->ii->gg;ff->ii;ff->gg");
-      const auto mb = gum::BayesNet< int >::fastPrototype(
+      const auto mb = gum::BayesNet< float >::fastPrototype(
          "cc->dd->ee;ff->dd->gg;ff->gg;ff->ii->gg");
       TS_ASSERT(gum::MarkovBlanket(bn, "dd").hasSameStructure(mb));
     }
